pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "gw2-compare"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "textual>=0.70",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
gw2-compare = "gw2compare.app:run"
